stuck on min/max

+++ db/migrate/201311152333_add_min_max_sponsor_to_buckets.rb
@@ -1,7 +1,7 @@
 class AddMinMaxSponsorToBuckets < ActiveRecord::Migration
   def change
-    add_column :buckets, :minimum_cents, :integer
-    add_column :buckets, :maximum_cents, :integer
+    add_column :buckets, :minimum_cents, :decimal
+    add_column :buckets, :maximum_cents, :decimal
     add_column :buckets, :sponsor_id, :integer
   end
 end
